refactor: ErrorBoundary로 감싸진 컴포넌트를 `AsyncBoundary`로 교체 #230
