+++ src/pages/study/assignment/[id].tsx
@@ -1,17 +1,13 @@
 import AssignmentList from '@components/assignment/AssignmentList';
-import SSRSafeSuspense from '@components/common/SSRSafeSuspense';
+import AsyncBoundary from '@components/async/AsyncBoundary';
 import MenuBar from '@components/common/MenuBar';
 import PageLayout from '@components/common/PageLayout';
-import ErrorBoundary from '@components/errorboundary/ErrorBoundary';
-import ErrorFallback from '@components/errorboundary/ErrorFallback';
 import AssignmentListSkeleton from '@components/skeleton/AssignmentListSkeleton';
-import { useQueryErrorResetBoundary } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 
 const Assignment = () => {
   const router = useRouter();
   const uuid = String(router.query.id);
-  const { reset } = useQueryErrorResetBoundary();
 
   return (
     <PageLayout>
@@ -25,11 +21,9 @@ const Assignment = () => {
           rightPath={`/study/member/${uuid}`}
         />
       </div>
-      <ErrorBoundary fallback={ErrorFallback} reset={reset}>
-        <SSRSafeSuspense fallback={<AssignmentListSkeleton />}>
-          <AssignmentList />
-        </SSRSafeSuspense>
-      </ErrorBoundary>
+      <AsyncBoundary suspenseFallback={<AssignmentListSkeleton />}>
+        <AssignmentList />
+      </AsyncBoundary>
     </PageLayout>
   );
 };
