+++ src/pages/study/assignment/detail/[...id].tsx
@@ -1,17 +1,13 @@
 import Loader from '@components/common/Loader';
 import MenuBar from '@components/common/MenuBar';
 import PageLayout from '@components/common/PageLayout';
-import SSRSafeSuspense from '@components/common/SSRSafeSuspense';
 import AssignmentDetailContent from '@components/detail/AssignmentDetailContent';
-import ErrorBoundary from '@components/errorboundary/ErrorBoundary';
-import ErrorFallback from '@components/errorboundary/ErrorFallback';
-import { useQueryErrorResetBoundary } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
+import AsyncBoundary from '@components/async/AsyncBoundary';
 
 const AssignmentDetail = () => {
   const router = useRouter();
   const [uuid, id] = (router.query.id as string[]) || [];
-  const { reset } = useQueryErrorResetBoundary();
 
   return (
     <PageLayout>
@@ -25,11 +21,9 @@ const AssignmentDetail = () => {
           centerPath={`/study/assignment/${uuid}`}
           rightPath={`/study/member/${uuid}`}
         />
-        <ErrorBoundary fallback={ErrorFallback} reset={reset}>
-          <SSRSafeSuspense fallback={<Loader isFull />}>
-            <AssignmentDetailContent />
-          </SSRSafeSuspense>
-        </ErrorBoundary>
+        <AsyncBoundary suspenseFallback={<Loader isFull />}>
+          <AssignmentDetailContent />
+        </AsyncBoundary>
       </div>
     </PageLayout>
   );
